% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{activate_gateway}
\alias{activate_gateway}
\title{Activates the gateway you previously deployed on your host}
\usage{
activate_gateway(
  ActivationKey = "string",
  GatewayName = "string",
  GatewayTimezone = "string",
  GatewayRegion = "string",
  GatewayType = "string",
  TapeDriveType = "string",
  MediumChangerType = "string"
)
}
\arguments{
\item{ActivationKey}{[required] Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter \code{activationKey}. It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the \code{ActivateGateway} API call determine the actual configuration of your gateway.

For more information, see https://docs.aws.amazon.com/storagegateway/latest/userguide/get-activation-key.html in the Storage Gateway User Guide.}

\item{GatewayName}{[required] The name you configured for your gateway.}

\item{GatewayTimezone}{[required] A value that indicates the time zone you want to set for the gateway. The time zone is of the format "GMT-hr:mm" or "GMT+hr:mm". For example, GMT-4:00 indicates the time is 4 hours behind GMT. GMT+2:00 indicates the time is 2 hours ahead of GMT. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.}

\item{GatewayRegion}{[required] A value that indicates the region where you want to store your data. The gateway region specified must be the same region as the region in your \code{Host} header in the request. For more information about available regions and endpoints for AWS Storage Gateway, see \href{http://docs.aws.amazon.com/general/latest/gr/rande.html#sg_region}{Regions and Endpoints} in the \emph{Amazon Web Services Glossary}.

Valid Values: "us-east-1", "us-east-2", "us-west-1", "us-west-2", "ca-central-1", "eu-west-1", "eu-central-1", "eu-west-2", "eu-west-3", "ap-northeast-1", "ap-northeast-2", "ap-southeast-1", "ap-southeast-2", "ap-south-1", "sa-east-1"}

\item{GatewayType}{A value that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is \code{CACHED}.

Valid Values: "STORED", "CACHED", "VTL", "FILE\_S3"}

\item{TapeDriveType}{The value that indicates the type of tape drive to use for tape gateway. This field is optional.

Valid Values: "IBM-ULT3580-TD5"}

\item{MediumChangerType}{The value that indicates the type of medium changer to use for tape gateway. This field is optional.

Valid Values: "STK-L700", "AWS-Gateway-VTL"}
}
\description{
Activates the gateway you previously deployed on your host. In the activation process, you specify information such as the region you want to use for storing snapshots or tapes, the time zone for scheduled snapshots the gateway snapshot schedule window, an activation key, and a name for your gateway. The activation process also associates your gateway with your account; for more information, see UpdateGatewayInformation.
}
\details{
You must turn on the gateway VM before you can activate your gateway.
}
\examples{
# Activates the gateway you previously deployed on your host.
activate_gateway(
  ActivationKey = "29AV1-3OFV9-VVIUB-NKT0I-LRO6V",
  GatewayName = "My_Gateway",
  GatewayRegion = "us-east-1",
  GatewayTimezone = "GMT-12:00",
  GatewayType = "STORED",
  MediumChangerType = "AWS-Gateway-VTL",
  TapeDriveType = "IBM-ULT3580-TD5"
)

}
