% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{add_cache}
\alias{add_cache}
\title{Configures one or more gateway local disks as cache for a gateway}
\usage{
add_cache(
  GatewayARN = "string",
  DiskIds = list(
    "string"
  )
)
}
\arguments{
\item{GatewayARN}{[required]}

\item{DiskIds}{[required]}
}
\description{
Configures one or more gateway local disks as cache for a gateway. This operation is only supported in the cached volume, tape and file gateway type (see \href{http://docs.aws.amazon.com/storagegateway/latest/userguide/StorageGatewayConcepts.html}{Storage Gateway Concepts}).
}
\details{
In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add cache, and one or more disk IDs that you want to configure as cache.
}
\examples{
# The following example shows a request that activates a gateway-stored
# volume.
add_cache(
  DiskIds = list(
    "pci-0000:03:00.0-scsi-0:0:0:0",
    "pci-0000:03:00.0-scsi-0:0:1:0"
  ),
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

}
