% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{add_upload_buffer}
\alias{add_upload_buffer}
\title{Configures one or more gateway local disks as upload buffer for a specified gateway}
\usage{
add_upload_buffer(GatewayARN, DiskIds)
}
\arguments{
\item{GatewayARN}{}

\item{DiskIds}{}
}
\description{
Configures one or more gateway local disks as upload buffer for a specified gateway. This operation is supported for the stored volume, cached volume and tape gateway types.
}
\details{
In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add upload buffer, and one or more disk IDs that you want to configure as upload buffer.
}
\examples{
# Configures one or more gateway local disks as upload buffer for a
# specified gateway.
add_upload_buffer(
  DiskIds = list(
    "pci-0000:03:00.0-scsi-0:0:0:0",
    "pci-0000:03:00.0-scsi-0:0:1:0"
  ),
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

}
