% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{cancel_retrieval}
\alias{cancel_retrieval}
\title{Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to a gateway after the retrieval process is initiated}
\usage{
cancel_retrieval(
  GatewayARN = "string",
  TapeARN = "string"
)
}
\arguments{
\item{GatewayARN}{[required]}

\item{TapeARN}{[required] The Amazon Resource Name (ARN) of the virtual tape you want to cancel retrieval for.}
}
\description{
Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to a gateway after the retrieval process is initiated. The virtual tape is returned to the VTS. This operation is only supported in the tape gateway type.
}
\examples{
# Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to
# a gateway after the retrieval process is initiated.
cancel_retrieval(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B",
  TapeARN = "arn:aws:storagegateway:us-east-1:999999999999:tape/AMZN01A2A4"
)

}
