% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{create_snapshot_from_volume_recovery_point}
\alias{create_snapshot_from_volume_recovery_point}
\title{Initiates a snapshot of a gateway from a volume recovery point}
\usage{
create_snapshot_from_volume_recovery_point(
  VolumeARN = "string",
  SnapshotDescription = "string"
)
}
\arguments{
\item{VolumeARN}{[required]}

\item{SnapshotDescription}{[required]}
}
\description{
Initiates a snapshot of a gateway from a volume recovery point. This operation is only supported in the cached volume gateway type.
}
\details{
A volume recovery point is a point in time at which all data of the volume is consistent and from which you can create a snapshot. To get a list of volume recovery point for cached volume gateway, use ListVolumeRecoveryPoints.

In the \code{CreateSnapshotFromVolumeRecoveryPoint} request, you identify the volume by providing its Amazon Resource Name (ARN). You must also provide a description for the snapshot. When the gateway takes a snapshot of the specified volume, the snapshot and its description appear in the AWS Storage Gateway console. In response, the gateway returns you a snapshot ID. You can use this snapshot ID to check the snapshot progress or later use it when you want to create a volume from a snapshot.

To list or delete a snapshot, you must use the Amazon EC2 API. For more information, in \emph{Amazon Elastic Compute Cloud API Reference}.
}
\examples{
# Initiates a snapshot of a gateway from a volume recovery point.
create_snapshot_from_volume_recovery_point(
  SnapshotDescription = "My root volume snapshot as of 2017-06-30T10:10:10.000Z",
  VolumeARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B/volume/vol-1122AABB"
)

}
