% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{create_storedi_scsi_volume}
\alias{create_storedi_scsi_volume}
\title{Creates a volume on a specified gateway}
\usage{
create_storedi_scsi_volume(GatewayARN, DiskId, PreserveExistingData, TargetName,
  NetworkInterfaceId, SnapshotId = NULL, KMSEncrypted = NULL,
  KMSKey = NULL)
}
\arguments{
\item{GatewayARN}{}

\item{DiskId}{The unique identifier for the gateway local disk that is configured as a stored volume. Use \href{http://docs.aws.amazon.com/storagegateway/latest/userguide/API_ListLocalDisks.html}{ListLocalDisks} to list disk IDs for a gateway.}

\item{PreserveExistingData}{Specify this field as true if you want to preserve the data on the local disk. Otherwise, specifying this field as false creates an empty volume.

Valid Values: true, false}

\item{TargetName}{The name of the iSCSI target used by initiators to connect to the target and as a suffix for the target ARN. For example, specifying \code{TargetName} as \emph{myvolume} results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gateway/sgw-12A3456B/target/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes of a gateway.}

\item{NetworkInterfaceId}{The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway.

Valid Values: A valid IP address.}

\item{SnapshotId}{The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot otherwise do not include this field. To list snapshots for your account use \href{http://docs.aws.amazon.com/AWSEC2/latest/APIReference/ApiReference-query-DescribeSnapshots.html}{DescribeSnapshots} in the \emph{Amazon Elastic Compute Cloud API Reference}.}

\item{KMSEncrypted}{True to use Amazon S3 server side encryption with your own AWS KMS key, or false to use a key managed by Amazon S3. Optional.}

\item{KMSKey}{The Amazon Resource Name (ARN) of the KMS key used for Amazon S3 server side encryption. This value can only be set when KMSEncrypted is true. Optional.}
}
\description{
Creates a volume on a specified gateway. This operation is only supported in the stored volume gateway type.
}
\details{
The size of the volume to create is inferred from the disk size. You can choose to preserve existing data on the disk, create volume from an existing snapshot, or create an empty volume. If you choose to create an empty gateway volume, then any existing data on the disk is erased.

In the request you must specify the gateway and the disk information on which you are creating the volume. In response, the gateway creates the volume and returns volume information such as the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target.
}
\examples{
# Creates a stored volume on a specified stored gateway.
create_storedi_scsi_volume(
  DiskId = "pci-0000:03:00.0-scsi-0:0:0:0",
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B",
  NetworkInterfaceId = "10.1.1.1",
  PreserveExistingData = TRUE,
  SnapshotId = "snap-f47b7b94",
  TargetName = "my-volume"
)

}
