% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{delete_file_share}
\alias{delete_file_share}
\title{Deletes a file share from a file gateway}
\usage{
delete_file_share(FileShareARN, ForceDelete = NULL)
}
\arguments{
\item{FileShareARN}{The Amazon Resource Name (ARN) of the file share to be deleted.}

\item{ForceDelete}{If this value is set to true, the operation deletes a file share immediately and aborts all data uploads to AWS. Otherwise, the file share is not deleted until all data is uploaded to AWS. This process aborts the data upload process, and the file share enters the FORCE\_DELETING status.}
}
\description{
Deletes a file share from a file gateway. This operation is only supported for file gateways.
}
\examples{

}
