% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{delete_snapshot_schedule}
\alias{delete_snapshot_schedule}
\title{Deletes a snapshot of a volume}
\usage{
delete_snapshot_schedule(VolumeARN)
}
\arguments{
\item{VolumeARN}{}
}
\description{
Deletes a snapshot of a volume.
}
\details{
You can take snapshots of your gateway volumes on a scheduled or ad hoc basis. This API action enables you to delete a snapshot schedule for a volume. For more information, see \href{http://docs.aws.amazon.com/storagegateway/latest/userguide/WorkingWithSnapshots.html}{Working with Snapshots}. In the \code{DeleteSnapshotSchedule} request, you identify the volume by providing its Amazon Resource Name (ARN). This operation is only supported in stored and cached volume gateway types.

To list or delete a snapshot, you must use the Amazon EC2 API. in \emph{Amazon Elastic Compute Cloud API Reference}.
}
\examples{
# This action enables you to delete a snapshot schedule for a volume.
delete_snapshot_schedule(
  VolumeARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B/volume/vol-1122AABB"
)

}
