% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{delete_tape_archive}
\alias{delete_tape_archive}
\title{Deletes the specified virtual tape from the virtual tape shelf (VTS)}
\usage{
delete_tape_archive(
  TapeARN = "string"
)
}
\arguments{
\item{TapeARN}{[required] The Amazon Resource Name (ARN) of the virtual tape to delete from the virtual tape shelf (VTS).}
}
\description{
Deletes the specified virtual tape from the virtual tape shelf (VTS). This operation is only supported in the tape gateway type.
}
\examples{
# Deletes the specified virtual tape from the virtual tape shelf (VTS).
delete_tape_archive(
  TapeARN = "arn:aws:storagegateway:us-east-1:204469490176:tape/TEST05A2A0"
)

}
