% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{describe_bandwidth_rate_limit}
\alias{describe_bandwidth_rate_limit}
\title{Returns the bandwidth rate limits of a gateway}
\usage{
describe_bandwidth_rate_limit(
  GatewayARN = "string"
)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Returns the bandwidth rate limits of a gateway. By default, these limits are not set, which means no bandwidth rate limiting is in effect.
}
\details{
This operation only returns a value for a bandwidth rate limit only if the limit is set. If no limits are set for the gateway, then this operation returns only the gateway ARN in the response body. To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.
}
\examples{
# Returns a value for a bandwidth rate limit if set. If not set, then only
# the gateway ARN is returned.
describe_bandwidth_rate_limit(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

}
