% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{describe_cache}
\alias{describe_cache}
\title{Returns information about the cache of a gateway}
\usage{
describe_cache(GatewayARN)
}
\arguments{
\item{GatewayARN}{}
}
\description{
Returns information about the cache of a gateway. This operation is only supported in the cached volume, tape and file gateway types.
}
\details{
The response includes disk IDs that are configured as cache, and it includes the amount of cache allocated and used.
}
\examples{
# Returns information about the cache of a gateway.
describe_cache(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

}
