% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{describe_gateway_information}
\alias{describe_gateway_information}
\title{Returns metadata about a gateway such as its name, network interfaces, configured time zone, and the state (whether the gateway is running or not)}
\usage{
describe_gateway_information(
  GatewayARN = "string"
)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Returns metadata about a gateway such as its name, network interfaces, configured time zone, and the state (whether the gateway is running or not). To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.
}
\examples{
# Returns metadata about a gateway such as its name, network interfaces,
# configured time zone, and the state (whether the gateway is running or
# not).
describe_gateway_information(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

}
