% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{describe_maintenance_start_time}
\alias{describe_maintenance_start_time}
\title{Returns your gateway's weekly maintenance start time including the day and time of the week}
\usage{
describe_maintenance_start_time(GatewayARN)
}
\arguments{
\item{GatewayARN}{}
}
\description{
Returns your gateway's weekly maintenance start time including the day and time of the week. Note that values are in terms of the gateway's time zone.
}
\examples{
# Returns your gateway's weekly maintenance start time including the day
# and time of the week.
describe_maintenance_start_time(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

}
