% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{describe_smb_file_shares}
\alias{describe_smb_file_shares}
\title{Gets a description for one or more Server Message Block (SMB) file shares from a file gateway}
\usage{
describe_smb_file_shares(FileShareARNList)
}
\arguments{
\item{FileShareARNList}{[required] An array containing the Amazon Resource Name (ARN) of each file share to be described.}
}
\description{
Gets a description for one or more Server Message Block (SMB) file shares from a file gateway. This operation is only supported for file gateways.
}
\section{Accepted Parameters}{
\preformatted{describe_smb_file_shares(
  FileShareARNList = list(
    "string"
  )
)
}
}

