% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{describe_smb_settings}
\alias{describe_smb_settings}
\title{Gets a description of a Server Message Block (SMB) file share settings from a file gateway}
\usage{
describe_smb_settings(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Gets a description of a Server Message Block (SMB) file share settings from a file gateway. This operation is only supported for file gateways.
}
\section{Accepted Parameters}{
\preformatted{describe_smb_settings(
  GatewayARN = "string"
)
}
}

