% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{describe_upload_buffer}
\alias{describe_upload_buffer}
\title{Returns information about the upload buffer of a gateway}
\usage{
describe_upload_buffer(
  GatewayARN = "string"
)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Returns information about the upload buffer of a gateway. This operation is supported for the stored volume, cached volume and tape gateway types.
}
\details{
The response includes disk IDs that are configured as upload buffer space, and it includes the amount of upload buffer space allocated and used.
}
\examples{
# Returns information about the upload buffer of a gateway including disk
# IDs and the amount of upload buffer space allocated/used.
describe_upload_buffer(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

# Returns information about the upload buffer of a gateway including disk
# IDs and the amount of upload buffer space allocated and used.
describe_upload_buffer(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

}
