% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{describe_vtl_devices}
\alias{describe_vtl_devices}
\title{Returns a description of virtual tape library (VTL) devices for the specified tape gateway}
\usage{
describe_vtl_devices(GatewayARN, VTLDeviceARNs = NULL, Marker = NULL,
  Limit = NULL)
}
\arguments{
\item{GatewayARN}{[required]}

\item{VTLDeviceARNs}{An array of strings, where each string represents the Amazon Resource Name (ARN) of a VTL device.

All of the specified VTL devices must be from the same gateway. If no VTL devices are specified, the result will contain all devices on the specified gateway.}

\item{Marker}{An opaque string that indicates the position at which to begin describing the VTL devices.}

\item{Limit}{Specifies that the number of VTL devices described be limited to the specified number.}
}
\description{
Returns a description of virtual tape library (VTL) devices for the specified tape gateway. In the response, AWS Storage Gateway returns VTL device information.
}
\details{
This operation is only supported in the tape gateway type.
}
\section{Accepted Parameters}{
\preformatted{describe_vtl_devices(
  GatewayARN = "string",
  VTLDeviceARNs = list(
    "string"
  ),
  Marker = "string",
  Limit = 123
)
}
}

\examples{
# Returns a description of virtual tape library (VTL) devices for the
# specified gateway.
\donttest{describe_vtl_devices(
  GatewayARN = "arn:aws:storagegateway:us-east-1:999999999999:gateway/sgw-12A3456B",
  Limit = 123L,
  Marker = "1",
  VTLDeviceARNs = list()
)}

}
