% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{describe_working_storage}
\alias{describe_working_storage}
\title{Returns information about the working storage of a gateway}
\usage{
describe_working_storage(GatewayARN)
}
\arguments{
\item{GatewayARN}{}
}
\description{
Returns information about the working storage of a gateway. This operation is only supported in the stored volumes gateway type. This operation is deprecated in cached volumes API version (20120630). Use DescribeUploadBuffer instead.
}
\details{
Working storage is also referred to as upload buffer. You can also use the DescribeUploadBuffer operation to add upload buffer to a stored volume gateway.

The response includes disk IDs that are configured as working storage, and it includes the amount of working storage allocated and used.
}
\examples{
# This operation is supported only for the gateway-stored volume
# architecture. This operation is deprecated in cached-volumes API version
# (20120630). Use DescribeUploadBuffer instead.
describe_working_storage(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

}
