% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{join_domain}
\alias{join_domain}
\title{Adds a file gateway to an Active Directory domain}
\usage{
join_domain(GatewayARN, DomainName, UserName, Password)
}
\arguments{
\item{GatewayARN}{[required] The unique Amazon Resource Name (ARN) of the file gateway you want to add to the Active Directory domain.}

\item{DomainName}{[required] The name of the domain that you want the gateway to join.}

\item{UserName}{[required] Sets the user name of user who has permission to add the gateway to the Active Directory domain.}

\item{Password}{[required] Sets the password of the user who has permission to add the gateway to the Active Directory domain.}
}
\description{
Adds a file gateway to an Active Directory domain. This operation is only supported for file gateways that support the SMB file protocol.
}
\section{Accepted Parameters}{
\preformatted{join_domain(
  GatewayARN = "string",
  DomainName = "string",
  UserName = "string",
  Password = "string"
)
}
}

