% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{list_volume_initiators}
\alias{list_volume_initiators}
\title{Lists iSCSI initiators that are connected to a volume}
\usage{
list_volume_initiators(VolumeARN)
}
\arguments{
\item{VolumeARN}{[required] The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes for the gateway.}
}
\description{
Lists iSCSI initiators that are connected to a volume. You can use this operation to determine whether a volume is being used or not. This operation is only supported in the cached volume and stored volume gateway types.
}
\section{Accepted Parameters}{
\preformatted{list_volume_initiators(
  VolumeARN = "string"
)
}
}

