% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{list_volume_recovery_points}
\alias{list_volume_recovery_points}
\title{Lists the recovery points for a specified gateway}
\usage{
list_volume_recovery_points(
  GatewayARN = "string"
)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Lists the recovery points for a specified gateway. This operation is only supported in the cached volume gateway type.
}
\details{
Each cache volume has one recovery point. A volume recovery point is a point in time at which all data of the volume is consistent and from which you can create a snapshot or clone a new cached volume from a source volume. To create a snapshot from a volume recovery point use the CreateSnapshotFromVolumeRecoveryPoint operation.
}
\examples{
# Lists the recovery points for a specified gateway in which all data of
# the volume is consistent and can be used to create a snapshot.
list_volume_recovery_points(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)

}
