% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{retrieve_tape_archive}
\alias{retrieve_tape_archive}
\title{Retrieves an archived virtual tape from the virtual tape shelf (VTS) to a tape gateway}
\usage{
retrieve_tape_archive(TapeARN, GatewayARN)
}
\arguments{
\item{TapeARN}{The Amazon Resource Name (ARN) of the virtual tape you want to retrieve from the virtual tape shelf (VTS).}

\item{GatewayARN}{The Amazon Resource Name (ARN) of the gateway you want to retrieve the virtual tape to. Use the ListGateways operation to return a list of gateways for your account and region.

You retrieve archived virtual tapes to only one gateway and the gateway must be a tape gateway.}
}
\description{
Retrieves an archived virtual tape from the virtual tape shelf (VTS) to a tape gateway. Virtual tapes archived in the VTS are not associated with any gateway. However after a tape is retrieved, it is associated with a gateway, even though it is also listed in the VTS, that is, archive. This operation is only supported in the tape gateway type.
}
\details{
Once a tape is successfully retrieved to a gateway, it cannot be retrieved again to another gateway. You must archive the tape again before you can retrieve it to another gateway. This operation is only supported in the tape gateway type.
}
\examples{
# Retrieves an archived virtual tape from the virtual tape shelf (VTS) to
# a gateway-VTL. Virtual tapes archived in the VTS are not associated with
# any gateway.
retrieve_tape_archive(
  GatewayARN = "arn:aws:storagegateway:us-east-1:999999999999:gateway/sgw-12A3456B",
  TapeARN = "arn:aws:storagegateway:us-east-1:999999999999:tape/TEST0AA2AF"
)

}
