% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{update_gateway_information}
\alias{update_gateway_information}
\title{Updates a gateway's metadata, which includes the gateway's name and time zone}
\usage{
update_gateway_information(
  GatewayARN = "string",
  GatewayName = "string",
  GatewayTimezone = "string"
)
}
\arguments{
\item{GatewayARN}{[required]}

\item{GatewayName}{}

\item{GatewayTimezone}{}
}
\description{
Updates a gateway's metadata, which includes the gateway's name and time zone. To specify which gateway to update, use the Amazon Resource Name (ARN) of the gateway in your request.
}
\details{
For Gateways activated after September 2, 2015, the gateway's ARN contains the gateway ID rather than the gateway name. However, changing the name of the gateway has no effect on the gateway's ARN.
}
\examples{
# Updates a gateway's metadata, which includes the gateway's name and time
# zone.
update_gateway_information(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B",
  GatewayName = "MyGateway2",
  GatewayTimezone = "GMT-12:00"
)

}
