% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.storagegateway_operations.R
\name{update_gateway_software_now}
\alias{update_gateway_software_now}
\title{Updates the gateway virtual machine (VM) software}
\usage{
update_gateway_software_now(GatewayARN)
}
\arguments{
\item{GatewayARN}{[required]}
}
\description{
Updates the gateway virtual machine (VM) software. The request immediately triggers the software update.
}
\details{
When you make this request, you get a \code{200 OK} success response immediately. However, it might take some time for the update to complete. You can call DescribeGatewayInformation to verify the gateway is in the \code{STATE_RUNNING} state.

A software update forces a system restart of your gateway. You can minimize the chance of any disruption to your applications by increasing your iSCSI Initiators' timeouts. For more information about increasing iSCSI Initiator timeouts for Windows and Linux, see \href{http://docs.aws.amazon.com/storagegateway/latest/userguide/ConfiguringiSCSIClientInitiatorWindowsClient.html#CustomizeWindowsiSCSISettings}{Customizing Your Windows iSCSI Settings} and \href{http://docs.aws.amazon.com/storagegateway/latest/userguide/ConfiguringiSCSIClientInitiatorRedHatClient.html#CustomizeLinuxiSCSISettings}{Customizing Your Linux iSCSI Settings}, respectively.
}
\section{Accepted Parameters}{
\preformatted{update_gateway_software_now(
  GatewayARN = "string"
)
}
}

\examples{
# Updates the gateway virtual machine (VM) software. The request
# immediately triggers the software update.
\donttest{update_gateway_software_now(
  GatewayARN = "arn:aws:storagegateway:us-east-1:111122223333:gateway/sgw-12A3456B"
)}

}
